+++ packages/core/src/components/buttons/buttons.tsx
@@ -1,13 +1,6 @@
 import React from 'react';
-import {
-	SpaceProps,
-	ColorProps,
-	TypographyProps,
-	LayoutProps,
-	Flex,
-} from '../globals/globals';
+import { SpaceProps } from '../globals/globals';
 import { useClassNames } from '../../hooks/use-class-names';
-import { ArrowRight } from './icons';
 import styles from './buttons.module.scss';
 
 export type ButtonProps = {
@@ -64,78 +57,5 @@ export const Button: React.FC<ButtonProps> = ({
 				</>
 			)}
 		</button>
-	);
-};
-
-export type LinkProps = {
-	cfg?: SpaceProps & ColorProps & TypographyProps & LayoutProps;
-	className?: string;
-	underline?: boolean;
-	testId?: string;
-	anchorTag?: boolean;
-	buttonAppearance?: boolean;
-	appearance?: 'primary' | 'outlined';
-	intent?: 'none' | 'success' | 'warning' | 'danger' | 'special';
-	btnSize?: 'small' | 'medium' | 'large';
-	arrowBtn?: boolean;
-	arrowColor?: ColorProps['fill'];
-	[key: string]: any;
-};
-export const Link: React.FC<LinkProps> = ({
-	cfg: globalStyles,
-	underline = false,
-	className,
-	testId,
-	anchorTag,
-	buttonAppearance,
-	appearance = 'primary',
-	intent = 'none',
-	btnSize = 'medium',
-	arrowBtn,
-	arrowColor = 'white',
-	children,
-	...props
-}) => {
-	const classNames = buttonAppearance
-		? useClassNames(globalStyles, [
-				styles.button,
-				styles[`appearance-${appearance}`],
-				styles[`intent-${intent}`],
-				styles[`size-${btnSize}`],
-				buttonAppearance && styles[`link-button-${btnSize}`],
-				arrowBtn && styles.linkArrowBtn,
-				className,
-		  ])
-		: useClassNames(globalStyles, [
-				styles.link,
-				{ [styles['link-underline']]: underline },
-				className,
-		  ]);
-
-	const genericProps = {
-		'data-testid': testId,
-		className: classNames,
-		...props,
-	};
-
-	const anchorProps = {
-		href: null,
-		...genericProps,
-	};
-
-	const buttonProps = {
-		type: 'button',
-		...genericProps,
-	};
-
-	return React.createElement(
-		anchorTag ? 'a' : 'button',
-		anchorTag ? anchorProps : buttonProps,
-		arrowBtn && (
-			<Flex>
-				<ArrowRight fill={arrowColor} />
-			</Flex>
-		),
-		children,
 	);
 };
