--- packages/core/src/components/buttons/link.mdx
@@ -1,153 +1,0 @@
----
-name: Link
-menu: Core
-route: /core/links
----
-
-import { Props } from 'docz';
-import { Playground } from '@playground';
-import { Link } from './buttons';
-
-# Link
-
-To trigger an operation.
-
-## When To Use
-
-A link means an operation (or a series of operations). Clicking a link will trigger corresponding business logic.
-
-## Usage
-
-```js
-import { Link } from '@tpr/core';
-```
-
-## Examples
-
-### Basic usage
-
-<Playground>
-	<Link>Click to go back</Link>
-</Playground>
-
-### Usage with props
-
-<Playground>
-	<Link cfg={{ mr: 3 }}>Home</Link>
-	<Link cfg={{ mr: 3 }}>Gallery</Link>
-	<Link cfg={{ mr: 3 }} underline={true}>
-		Log out
-	</Link>
-	<Link cfg={{ mr: 3 }} disabled>
-		Disabled Link
-	</Link>
-</Playground>
-
-### Usage with colours from the theme
-
-<Playground>
-	<Link cfg={{ mr: 3, color: 'danger.2' }}>Home</Link>
-	<Link cfg={{ mr: 3, color: 'warning.1' }}>Gallery</Link>
-	<Link cfg={{ mr: 3, color: 'success.1' }}>Log out</Link>
-</Playground>
-
-### Anchor tag
-
-Using the anchorTag prop will create an `<a>` element, therefore the "disabled" prop will have no offect
-
-<Playground>
-	and their <Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		href="http://www.google.com"
-		target="_blank"
-	>
-		advisers
-	</Link>to identify the quality of the scheme's assets
-</Playground>
-
-### Anchor tag with button appearance
-
-<Playground>
-	<Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		buttonAppearance={true}
-		btnSize="large"
-		href="http://www.google.com"
-		target="_blank"
-	>
-		Home
-	</Link>
-	<Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		buttonAppearance={true}
-		btnSize="medium"
-		intent="warning"
-		href="http://www.google.com"
-		target="_blank"
-	>
-		Home
-	</Link>
-	<Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		buttonAppearance={true}
-		btnSize="small"
-		intent="danger"
-		href="http://www.google.com"
-		target="_blank"
-	>
-		Home
-	</Link>
-</Playground>
-
-### Anchor tag with ArrowButton appearance
-
-<Playground>
-	<Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		buttonAppearance={true}
-		btnSize="large"
-		href="http://www.google.com"
-		target="_blank"
-		arrowBtn={true}
-	>
-		Home
-	</Link>
-	<Link
-		cfg={{ mr: 3, color: 'black' }}
-		anchorTag={true}
-		buttonAppearance={true}
-		arrowBtn={true}
-		arrowColor="black"
-		btnSize="medium"
-		intent="warning"
-		href="http://www.google.com"
-		target="_blank"
-	>
-		Home
-	</Link>
-	<Link
-		cfg={{ mr: 3 }}
-		anchorTag={true}
-		buttonAppearance={true}
-		arrowBtn={true}
-		btnSize="small"
-		intent="danger"
-		href="http://www.google.com"
-		target="_blank"
-	>
-		Home
-	</Link>
-</Playground>
-
-## API
-
-```
-Accepted config props: SpaceProps, ColorProps, TypographyProps, LayoutProps
-```
-
-<Props of={Link} />
