+++ packages/core/src/components/links/links.module.scss
@@ -1,0 +1,59 @@
+@import '@tpr/theming/lib/variables.scss';
+
+.link {
+	display: inline-block;
+	background: transparent;
+	border: none;
+	color: $colors-primary-3;
+	font-family: $fonts-sans-serif;
+	font-size: $font-size-2;
+	font-weight: $font-weight-3;
+	text-decoration: underline;
+	cursor: pointer;
+
+	&:hover {
+		color: $colors-neutral-a1;
+	}
+	
+	&:disabled {
+		color: $colors-neutral-6;
+		text-decoration: none;
+		cursor: not-allowed;
+	}
+	
+	&:visited {
+		color: $colors-accents-2;
+		background-color: $colors-warning-1;
+	}
+	
+	&:focus {
+		color: $colors-neutral-8;
+		background: $colors-warning-1;
+		outline: none;
+		text-decoration: none;
+		box-shadow: 0px -2px $colors-warning-1, 0px 4px $colors-neutral-8;
+	}
+}
+
+.link-underline {
+	text-decoration: underline;
+}
+
+.link-button-small {
+	line-height: 40px;
+}
+
+.link-button-medium {
+	line-height: 50px;
+}
+
+.link-button-large {
+	line-height: 60px;
+}
+
+.linkArrowBtn {
+	display: inline-flex;
+	flex-direction: row-reverse;
+	justify-content: center;
+	align-items: center;
+}
