+++ packages/core/src/index.ts
@@ -4,4 +4,5 @@ export * from './utils';
 export * from './components/globals/globals';
 export * from './components/buttons/buttons';
 export * from './components/layout/layout';
+export * from './components/links/links';
 export * from './components/typography/typography';
